src/productpage/requirements.txt
requests
flask==0.12.2
flask_json
flask_bootstrap
json2html
simplejson
gevent
grpcio
grpcio-tools
